requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.3
